requirements.txt
aiogram==2.25.1
asyncpg==0.28.0
python-dotenv==1.0.0
aiohttp==3.8.5
openai==0.28.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
pandas==2.0.3
reportlab==4.0.4